Backend/image-service/requirements.txt
flux
Flask
psycopg2-binary
python-dotenv
diffusers
transformers
torch
accelerate
google-generativeai
uuid